Disabled caching for recent `playerRollDice` tx
This could get annoying if the user refreshed the roll history
after a long time with no bet and they placed a bet.
The refresh would retrieve not so high block number and that would
be cached for 5 minutes. That makes the user experience frustrating
because the user pull frequently to have bet results and doesn't
get it, even though it's been resolved.