Notifies user when roll processed, fixes #57
Adds a background service for pulling rolls and notify user when roll
was processed.
There's currently a bug where the service may die for a unknown reason.
Also the notification is not clickable and doesn't bring back to the
app.
This commit also brings a bunch of other things such as:

  - fixes missing `libffi-dev`
  - adds missing `libgmp3-dev` required by coincurve
  - upgrades Docker base image to latest Ubuntu Bionic LTS
  - refactors some file locations (shared with service/)