Shuts down service after inactivity
If there's no roll activity, the service will stop the pulling loop.
This is to save smartphone resources (bandwidth, CPU, battery...).
Placing another bet will restart the service if it's down.
However if the service is still up it will not restart.
This could be an issue since the activity counter won't be reset.
In another design we could use interprocess commmunication to update the
activity counter or restart the service if down.